Fix type_union_implicit rule for Draft 0, 1, and 2
The rule was checking for the wrong vocabulary identifiers (hyper-schema#
instead of schema#) when determining if enum is present for Draft 0, 1,
and 2. This caused the rule to incorrectly add type: [all types] even
when enum was already present.

Fixed the vocabulary checks from:
- http://json-schema.org/draft-02/hyper-schema#
- http://json-schema.org/draft-01/hyper-schema#
- http://json-schema.org/draft-00/hyper-schema#

To:
- http://json-schema.org/draft-02/schema#
- http://json-schema.org/draft-01/schema#
- http://json-schema.org/draft-00/schema#

Updated 6 test cases with correct expected values (without the type array):
- AlterSchema_lint_draft1.type_boolean_as_enum_1
- AlterSchema_lint_draft1.type_null_as_enum_1
- AlterSchema_lint_draft1.equal_numeric_bounds_to_enum_2
- AlterSchema_lint_draft2.type_boolean_as_enum_1
- AlterSchema_lint_draft2.type_null_as_enum_1
- AlterSchema_lint_draft2.equal_numeric_bounds_to_enum_2

Co-Authored-By: unknown <>