Fix type_union_implicit rule for Draft 0/1/2 to check schema# instead of hyper-schema#
The rule was incorrectly checking for hyper-schema# vocabulary URIs for
Draft 0, 1, and 2, but the actual schemas use schema# URIs. This caused
the rule to not properly skip adding the type array when enum was present.

Updated the vocabulary checks to use:
- http://json-schema.org/draft-00/schema#
- http://json-schema.org/draft-01/schema#
- http://json-schema.org/draft-02/schema#

Also updated 6 test cases to expect the correct behavior (enum without type array):
- AlterSchema_lint_draft1.type_boolean_as_enum_1
- AlterSchema_lint_draft1.type_null_as_enum_1
- AlterSchema_lint_draft1.equal_numeric_bounds_to_enum_2
- AlterSchema_lint_draft2.type_boolean_as_enum_1
- AlterSchema_lint_draft2.type_null_as_enum_1
- AlterSchema_lint_draft2.equal_numeric_bounds_to_enum_2

Co-Authored-By: unknown <>