Fix type_union_implicit rule to check enum for Draft 0/1/2 schema vocabularies
The rule was only checking for enum presence in the hyper-schema vocabularies
(draft-00/01/02-hyper-schema#) but not in the regular schema vocabularies
(draft-00/01/02/schema#). This caused the rule to incorrectly add type arrays
when enum was present in Draft 0/1/2 schemas.

Added checks for:
- http://json-schema.org/draft-02/schema#
- http://json-schema.org/draft-01/schema#
- http://json-schema.org/draft-00/schema#

Updated test cases to reflect the correct expected behavior where type should
not be added alongside enum.

Co-Authored-By: unknown <>